restructuring the old codes into corresponding components and importing them to app.jsx
